+++ packages/gasket-log/CHANGELOG.md
@@ -1,5 +1,9 @@
 # `@gasket/log`
 
+### 6.30.0
+
+- Support enabling logging in production ([#402])
+
 ### 6.1.0
 
 - Support for custom log levels ([#279])
@@ -63,3 +67,4 @@
 
 [#247]: https://github.com/godaddy/gasket/pull/247
 [#279]: https://github.com/godaddy/gasket/pull/279
+[#402]: https://github.com/godaddy/gasket/pull/402