+++ packages/gasket-plugin-elastic-apm/CHANGELOG.md
@@ -1,5 +1,9 @@
 # `@gasket/plugin-elastic-apm`
 
+### 6.30.0
+
+- Add an `apmTransaction` lifecycle ([#400])
+
 ### 6.27.1
 
 - Skip preboot lifecycle when app is run locally ([#388])
@@ -20,3 +24,4 @@
 [#311]:https://github.com/godaddy/gasket/pull/311
 [#378]:https://github.com/godaddy/gasket/pull/378
 [#388]:https://github.com/godaddy/gasket/pull/388
+[#400]:https://github.com/godaddy/gasket/pull/400